Add automatic dependency installation for 'make test'

+++ srcs/server/Server.cpp
@@ -6,7 +6,7 @@
 /*   By: gdosch <gdosch@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/16 10:19:49 by eschwart          #+#    #+#             */
-/*   Updated: 2026/02/05 13:12:02 by gdosch           ###   ########.fr       */
+/*   Updated: 2026/02/08 17:10:56 by gdosch           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -351,7 +351,7 @@ void Server::handleClientRead(size_t clientIndex)
 		{
 			// Start CGI asynchronously
 			const ServerConfig *config = selectConfig(client.getRequest(), clientFd);
-			size_t cgiExecutionTimeout = config ? config->getCgiTimeout() : DEFAULT_CGI_EXECUTION_TIMEOUT;
+			size_t cgiExecutionTimeout = config ? config->getCgiTimeout() : static_cast<size_t>(DEFAULT_CGI_EXECUTION_TIMEOUT);
 
 			// Store timing info for CGI logging
 			if (config)
